5-search engine/requirements.txt
langchain
langchain-openai
python-dotenv
langchain-community
ipykernel
langchain-core
streamlit
langchain-groq
langchain_huggingface
langchain_chroma
langchain_classic
pypdf
arxiv
wikipedia
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
pytube
langchain_yt_dlp
numexpr
huggingface_hub
cassio
datasets
openai
PyPDF2
tiktoken
langchain_astradb
gradio
requests



